% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dp.R
\name{core.dp}
\alias{core.dp}
\title{A}
\usage{
core.dp(infla, sub, pesos, janela = 48)
}
\arguments{
\item{infla}{Índice de inflação (variação).}

\item{sub}{Subitens em variação.}

\item{pesos}{Pesos correspondentes a cada subitem.}

\item{janela}{Janela para o cálculo da volatilidade}
}
\description{
A
}
\examples{

}
\keyword{core}
