% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dw.R
\name{core.dw}
\alias{core.dw}
\title{Computes the double weighting core inflation}
\usage{
core.dw(infla, sub, weights, wind = 48)
}
\arguments{
\item{infla}{A \code{ts} object. The inflation index (variation).}

\item{sub}{A \code{ts}. Subitems' variation.}

\item{weights}{A \code{ts}. Corresponding weights to each subitem.}

\item{wind}{An \code{integer}. The volatility's window to be computed.}
}
\value{
A \code{ts} object.
}
\examples{
\dontrun{
 ipca <- ipca_get(group = "subitem")
 nuc <- core.dw(ipca$ipca_index, ipca$ipca_ts, ipca$weights_ts, wind = 12)

 ipca <- ipca_get(group = "item")
 nuc <- core.dw(ipca$ipca_index, ipca$ipca_ts, ipca$weights_ts, wind = 24)

}
}
\keyword{core}
\keyword{weight}
