% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ex.R
\encoding{utf8}
\name{core.ex}
\alias{core.ex}
\title{Núcleo de inflação por exclusão}
\usage{
core.ex(sub, pesos, codigos, cortes = 4, alpha = 2)
}
\arguments{
\item{sub}{Subitens em variação.}

\item{pesos}{Pesos correspondentes a cada subitem.}

\item{codigos}{Codigos dos subitens}

\item{cortes}{Quantidade deseja de cortes a serem realizados na janela temporal.}

\item{alpha}{Nível de significância.}
}
\description{
Núcleo de inflação por exclusão
}
\examples{
ipc.ex1 <- core.ex(variacao.ts$subitens, pesos.ts$subitens)

# ex2: excluindo alimentação
novos_pesos <- pesos.ts$subitens
novos_pesos[,substr(colnames(novos_pesos), 5,5) == 1] <- 0
novos_pesos <- (novos_pesos/rowSums(novos_pesos, na.rm = TRUE))*100
ipc.ex3 <- ts(rowSums(variacao.ts$subitens*novos_pesos, 
na.rm = TRUE)/100, start = start(variacao.ts$subitens), frequency = 12)


# gráficos
ts.plot(variacao.ts$ipc, ipc.ex1, col = c(1:2), lwd = 1:2)
ts.plot(variacao.ts$ipc, ipc.ex3, col = c(1:2), lwd = 1:2)
}
\keyword{core}
