% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ma.R
\name{core.ma}
\alias{core.ma}
\title{A}
\usage{
core.ma(sub, pesos, inf = 20, sup = 20, suave = FALSE, janela = 12)
}
\arguments{
\item{sub}{Subitens em variação.}

\item{pesos}{Pesos correspondentes a cada subitem.}

\item{inf}{Corte percentual da cauda inferior. Predefinido como 20.}

\item{sup}{Corte percentual da cauda superior. Predefinido como 20.}

\item{suave}{Código dos itens que serão suavizados.}

\item{janela}{Tamanho da janela temporal de corte.}
}
\description{
A
}
\examples{

}
\keyword{core}
\keyword{ma}
\keyword{média}
\keyword{móvel}
