% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom3.R
\name{geom3}
\alias{geom3}
\title{Accumulate (3 months) and annualize}
\usage{
geom3(x, anual = F)
}
\arguments{
\item{x}{Time series}

\item{anual}{Should the series be annualized?}
}
\description{
!! DESCREVER
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}, Fernando Teixeira \email{fernando.teixeira@fgv.br}, Daiane Marcolino \email{daiane.marcolino@fgv.br}
}
