% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFL.core_dw.R
\name{INFL.core_dw}
\alias{INFL.core_dw}
\title{Computes the double weighting core inflation}
\usage{
INFL.core_dw(infl.var, subits.var, weights, wind = 48)
}
\arguments{
\item{infl.var}{A \code{ts} object. The inflation index variation.}

\item{subits.var}{A \code{ts}. Subitems' variation.}

\item{weights}{A \code{ts}. Weights corresponding to each subitem.}

\item{wind}{An \code{integer}. The volatility's window size to be computed.}
}
\value{
A \code{ts} object.
}
\keyword{core}
\keyword{weight}
