% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFL.core_ex.R
\encoding{utf8}
\name{INFL.core_ex}
\alias{INFL.core_ex}
\title{Computes the core inflation using the subitem exclusion method}
\usage{
INFL.core_ex(subitems, weights, info, n.blocks = 4, alpha = 2)
}
\arguments{
\item{subitems}{A \code{ts}. Subitems' variation.}

\item{weights}{A \code{ts}. Each subitem corresponding weights. If missing, all items get the same weight.}

\item{info}{A \code{data.frame}. Subitem metadata table containing their codes and descriptions.}

\item{alpha}{An \code{integer}. Significance level in percentage.}

\item{part}{An \code{integer}. Partitions' number inside the temporal window.}
}
\description{
Computes the core inflation using the subitem exclusion method
}
\keyword{core}
\keyword{exclusion}
