% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFLATION.R
\name{INFLATION}
\alias{INFLATION}
\title{An R package providing tools for those who want to figure out the core inflation of their series.}
\description{
The INFLATION package...
}
\keyword{bacen}
\keyword{bcb}
\keyword{core}
\keyword{economics}
\keyword{inflation}
