% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipca_get.R
\encoding{utf8}
\name{ipca_get}
\alias{ipca_get}
\title{Núcleo de inflação por exclusão}
\usage{
ipca_get(group = c("subgroup", "item", "subitem"))
}
\arguments{
\item{grupamento}{Grupamento determina o que será retornado: subgrupos, itens ou subitens}
}
\description{
Núcleo de inflação por exclusão
}
\examples{
ipc.ex1 <- core.ex(variacao.ts$subitens, pesos.ts$subitens)

# ex2: excluindo alimentação
novos_pesos <- pesos.ts$subitens
novos_pesos[,substr(colnames(novos_pesos), 5,5) == 1] <- 0
novos_pesos <- (novos_pesos/rowSums(novos_pesos, na.rm = TRUE))*100
ipc.ex3 <- ts(rowSums(variacao.ts$subitens*novos_pesos,
na.rm = TRUE)/100, start = start(variacao.ts$subitens), freq = 12)


# gráficos
ts.plot(variacao.ts$ipc, ipc.ex1, col = c(1:2), lwd = 1:2)
ts.plot(variacao.ts$ipc, ipc.ex3, col = c(1:2), lwd = 1:2)
}
\keyword{core}
\keyword{ipca}
