% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipca_item}
\alias{ipca_item}
\title{IPCA items and its weights}
\format{A list with five attributes:
\describe{
  \item{ipca}{dataframe with ipca items}
  \item{weights}{dataframe with weights items}
  \item{ipca_ts}{ts with ipca items}
  \item{weights_ts}{ts with weights items}
  \item{cod}{Items' codes}
}}
\source{
\url{https://sidra.ibge.gov.br}
}
\usage{
ipca_item
}
\description{
A dataset containing the IPCA items, their respective weights and codes in tibble format. Items and codes are also provided in ts data structure.
}
\keyword{datasets}
