% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipca_sub}
\alias{ipca_sub}
\title{IPCA subitems and its weights}
\format{A list with six attributes:
\describe{
  \item{ipca}{dataframe with ipca subitems}
  \item{weights}{dataframe with weights subitems}
  \item{ipca_ts}{ts with ipca subitems}
  \item{weights_ts}{ts with weights subitems}
  \item{cod}{Subitems' codes}
  \item{ipca_index}{The full index}
}}
\source{
\url{https://sidra.ibge.gov.br}
}
\usage{
ipca_sub
}
\description{
A dataset containing the IPCA items, their respective weights and codes in tibble format. Subitems and codes are also provided in ts data structure.
}
\keyword{datasets}
