% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol.mat.R
\encoding{utf8}
\name{vol.mat}
\alias{vol.mat}
\title{Computes the volatility matrix}
\usage{
vol.mat(x, info, n.blocks, alpha)
}
\arguments{
\item{x}{Subitems' variation.}

\item{info}{Subitems' metadata.}

\item{n.blocks}{Number of cuts to be made.}

\item{alpha}{Significance level.}
}
\description{
!! DESCREVER O QUE É A MATRIZ
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}, Fernando Teixeira \email{fernando.teixeira@fgv.br}, Daiane Marcolino \email{daiane.marcolino@fgv.br}
}
