% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ex.R
\encoding{utf8}
\name{vola}
\alias{vola}
\title{Computes the volatility in the core by exclusion}
\usage{
vola(x, codes, cortes, alpha)
}
\arguments{
\item{x}{Subitems' variation.}

\item{codes}{Subitems' codes.}

\item{cortes}{Number of cuts to be made.}

\item{alpha}{Significance level.}
}
\description{
Computes the volatility in the core by exclusion
}
