% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ex.R
\encoding{utf8}
\name{core.ex}
\alias{core.ex}
\title{Computes the core inflation using the subitem exclusion method}
\usage{
core.ex(sub, weights, codes, part = 4, alpha = 2)
}
\arguments{
\item{sub}{A \code{ts}. Subitems' variation.}

\item{weights}{A \code{ts}. Each subitem corresponding weights. If missing, all items get the
same weight.}

\item{codes}{A \code{data.frame}. Subitem containing the codes and their descriptions.}

\item{part}{An \code{integer}. Partitions' number inside the temporal window.}

\item{alpha}{An \code{integer}. Significance level in percentage.}
}
\description{
Computes the core inflation using the subitem exclusion method
}
\examples{
\dontrun{
load("~/INFLATION/data/codigos.rda")
load("~/INFLATION/data/pesos.ts.rda")
load("~/INFLATION/data/variacao.ts.rda")
codes <- codigos[2:nrow(codigos),]

ipc.ex1 <- core.ex(sub = variacao.ts$subitens, weights = pesos.ts$subitens, codes = codigos,
part = 4, alpha = 2)

# ex2: excluindo alimentação
novos_pesos <- pesos.ts$subitens
novos_pesos[,substr(colnames(novos_pesos), 5,5) == 1] <- 0
novos_pesos <- (novos_pesos/rowSums(novos_pesos, na.rm = TRUE))*100
ipc.ex3 <- ts(rowSums(variacao.ts$subitens*novos_pesos,
na.rm = TRUE)/100, start = start(variacao.ts$subitens), frequency = 12)


# gráficos
ts.plot(variacao.ts$ipc, ipc.ex1, col = c(1:2), lwd = 1:2)
ts.plot(variacao.ts$ipc, ipc.ex3, col = c(1:2), lwd = 1:2)

ipca <- ipca_get(group = "subitem")
# Estes códigos ainda são do IPC - mudar para IPCA
load("~/INFLATION/data/codigos.rda")
codigos <- codigos[1:373,]
nuc <- core.ex(sub = ipca$ipca_ts, weights = ipca$weights_ts, codes = ipca$cod)

}
}
\keyword{core}
\keyword{exclusion}
