% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFLATION.R
\docType{package}
\name{Inflation}
\alias{Inflation}
\alias{Inflation-package}
\title{An R package providing tools for those who want to figure out the core inflation of their series.}
\description{
The Inflation package provides tools that allow its user to better understand core inflation.

The package provides a set of functions that compute the core inflation based on items that are part of
an inflation index. Currently, the package allows for four different core inflation computations
methods: trimmed means, exclusion, double weighting and triple filter. The first three are well known by the
public. The latter is a method we developed that we believe is a better measure.
}
\note{
The authors would like to thank the support by the Getulio
Vargas Foundation (FGV) and make it clear that all
data in the package is in public domain. We reaffirm that Inflation
is mainly intended for academic usage.
}
\author{
Pedro Costa Ferreira \email{pedro.guilherme@fgv.br},
        Talitha Speranza \email{talitha.speranza@fgv.br},
        Fernando Teixeira \email{fernando.teixeira@fgv.br},
        Daiane Marcolino \email{daiane.mattos@fgv.br}
}
\keyword{bacen}
\keyword{bcb}
\keyword{core}
\keyword{economics}
\keyword{inflation}
