% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vola.R
\encoding{utf8}
\name{vola}
\alias{vola}
\title{Calcula a volatilidade para núcleo por exclusão}
\usage{
vola(x, codes, cortes, alpha)
}
\arguments{
\item{x}{Subitens em variação.}

\item{codes}{codes dos subitens}

\item{cortes}{Quantidade deseja de cortes a serem realizados na janela temporal.}

\item{alpha}{Nível de significância.}
}
\description{
Calcula a volatilidade para núcleo por exclusão
}
